added time to read to views

+++ src/Services/IPostInfo.cs
@@ -2,6 +2,6 @@
 {
     public interface IPostInfo
     {
-        double GetTimeToRead(string postHtml);
+        int GetTimeToRead(string postHtml);
     }
 }
